refactor : 면접 도메인 테이블 컬럼 수정 및 api요청 수정 / feat : 면접 담당자 도메인 Command/Query 기능 구현

+++ src/main/java/com/piveguyz/empickbackend/employment/interviews/facade/InterviewFacade.java
@@ -5,29 +5,41 @@ import com.piveguyz.empickbackend.common.response.ResponseCode;
 import com.piveguyz.empickbackend.employment.interviews.interview.command.application.dto.InterviewCommandDTO;
 import com.piveguyz.empickbackend.employment.interviews.interview.command.application.service.InterviewCommandService;
 import com.piveguyz.empickbackend.employment.interviews.interview.query.service.InterviewQueryService;
+import com.piveguyz.empickbackend.employment.interviews.interviewScore.command.application.dto.InterviewScoreCommandDTO;
 import com.piveguyz.empickbackend.employment.interviews.interviewScore.command.application.service.InterviewScoreCommandService;
+import com.piveguyz.empickbackend.employment.interviews.interviewScore.command.domain.aggregate.InterviewScoreEntity;
+import com.piveguyz.empickbackend.employment.interviews.interviewScore.command.domain.repository.InterviewScoreRepository;
 import com.piveguyz.empickbackend.employment.interviews.interviewScore.query.dto.InterviewScoreQueryDTO;
 import com.piveguyz.empickbackend.employment.interviews.interviewScore.query.service.InterviewScoreQueryService;
+import com.piveguyz.empickbackend.employment.interviews.interviewSheetItem.command.application.dto.InterviewSheetItemCommandDTO;
+import com.piveguyz.empickbackend.employment.interviews.interviewSheetItem.command.domain.repository.InterviewSheetItemRepository;
+import com.piveguyz.empickbackend.employment.interviews.interviewSheetItem.query.dto.InterviewSheetItemQueryDTO;
+import com.piveguyz.empickbackend.employment.interviews.interviewSheetItem.query.service.InterviewSheetItemQueryService;
+import com.piveguyz.empickbackend.employment.interviews.interviewer.command.application.dto.InterviewerCommandDTO;
+import com.piveguyz.empickbackend.employment.interviews.interviewer.command.application.mapper.InterviewerCommandMapper;
+import com.piveguyz.empickbackend.employment.interviews.interviewer.command.application.service.InterviewerCommandService;
+import com.piveguyz.empickbackend.employment.interviews.interviewer.command.domain.aggregate.InterviewerEntity;
+import com.piveguyz.empickbackend.employment.interviews.interviewer.command.domain.repository.InterviewerRepository;
+import com.piveguyz.empickbackend.employment.interviews.interviewer.query.dto.InterviewerQueryDTO;
+import com.piveguyz.empickbackend.employment.interviews.interviewer.query.mapper.InterviewerMapper;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
+@AllArgsConstructor
 @Service
 public class InterviewFacade {
-    private final InterviewScoreCommandService interviewScoreCommandService;
-    private final InterviewScoreQueryService interviewScoreQueryService;
     private final InterviewCommandService interviewCommandService;
+    private final InterviewScoreQueryService interviewScoreQueryService;
     private final InterviewQueryService interviewQueryService;
-
-    public InterviewFacade(InterviewScoreQueryService interviewScoreQueryService,
-                           InterviewScoreCommandService interviewScoreCommandService,
-                           InterviewCommandService interviewCommandService, InterviewQueryService interviewQueryService) {
-        this.interviewScoreCommandService = interviewScoreCommandService;
-        this.interviewScoreQueryService = interviewScoreQueryService;
-        this.interviewCommandService = interviewCommandService;
-        this.interviewQueryService = interviewQueryService;
-    }
+    private final InterviewSheetItemQueryService interviewSheetItemQueryService;
+    private final InterviewerRepository interviewerRepository;
+    private final InterviewerCommandMapper interviewerCommandMapper;
+    private final InterviewScoreCommandService interviewScoreCommandService;
+    private final InterviewerCommandService interviewerCommandService;
+    private final InterviewerMapper interviewerMapper;
 
     public InterviewCommandDTO createInterview(InterviewCommandDTO interviewCommandDTO){
         LocalDateTime interviewTime = interviewCommandDTO.getDatetime();
@@ -41,12 +53,41 @@ public class InterviewFacade {
     }
 
     public InterviewCommandDTO deleteInterview(Integer id){
-        List<InterviewScoreQueryDTO> interviewScoreQueryDTOList = interviewScoreQueryService.findByInterviewId(id);
-        for(InterviewScoreQueryDTO interviewScoreQueryDTO : interviewScoreQueryDTOList){
-            Integer interviewScoreId = interviewScoreQueryDTO.getId();
-            interviewScoreCommandService.delete(interviewScoreId);
+        List<InterviewerQueryDTO> interviewerQueryDTOList = interviewerMapper.findByInterviewId(id);
+        if(!interviewerQueryDTOList.isEmpty()){
+            throw new BusinessException(ResponseCode.BAD_REQUEST);
         }
         InterviewCommandDTO dto = interviewCommandService.delete(id);
         return dto;
+    }
+
+    public InterviewerCommandDTO updateInterviewerScore(Integer id) {
+        InterviewerEntity entity = interviewerRepository.findById(id)
+                .orElseThrow(() -> new BusinessException(ResponseCode.NOT_FOUND));
+        Double totalScore = (double) 0;
+        List<InterviewScoreQueryDTO> dtoList = interviewScoreQueryService.findByInterviewerId(id);
+        for(InterviewScoreQueryDTO scoreDTO : dtoList){
+            Integer score = scoreDTO.getScore();
+            Integer itemId = scoreDTO.getItemId();
+            InterviewSheetItemQueryDTO itemDTO = interviewSheetItemQueryService.findById(itemId);
+            Double weight = itemDTO.getWeight();
+            totalScore += score * weight;
+        }
+        entity.setScore(totalScore);
+        interviewerRepository.save(entity);
+        return interviewerCommandMapper.toDTO(entity);
+    }
+
+    public InterviewerCommandDTO deleteInterviewer(Integer id) {
+        List<InterviewScoreQueryDTO> dtoList = interviewScoreQueryService.findByInterviewerId(id);
+        if(!dtoList.isEmpty()){
+            throw new BusinessException(ResponseCode.BAD_REQUEST);
+        }
+//        for(InterviewScoreQueryDTO scoreDTO : dtoList){
+//            Integer scoreId = scoreDTO.getId();
+//            interviewScoreCommandService.delete(scoreId);
+//        }
+        InterviewerCommandDTO deletedDTO = interviewerCommandService.deleteInterviewer(id);
+        return deletedDTO;
     }
 }
